refactor: 최신버전으로 병합 및 spinner파일 위치옮김
