+++ src/components/common/Button.jsx
@@ -7,6 +7,7 @@ const Button = ({
   background = "linear-gradient(90deg, #f86f65 0%, $color-brand-pink 100%)",
   onClick,
   children,
+  borderRadius = "3px",
   className, // className을 props로 받아서
 }) => {
   return (
@@ -17,6 +18,7 @@ const Button = ({
         width,
         color,
         background,
+        borderRadius,
       }}
       className={className} // button과 전달받은 className 적용
     >
