+++ src/components/slider/slider.scss
@@ -23,6 +23,7 @@
 
 .sliderBox {
   width: 1165px;
+  height: 407.57px;
 }
 .tributeSupport {
   padding: 20px 9px;
@@ -129,6 +130,7 @@
 @media (max-width: 744px) {
   .sliderBox {
     width: 744px;
+    height: 463.71px;
   }
   .tributeSupport {
     padding: 20px 9px;
@@ -142,6 +144,7 @@
 @media (max-width: 374px) {
   .sliderBox {
     width: 374px;
+    height: 303px;
   }
   .tributeSupport {
     padding: 20px 9px;
