+++ src/pages/ListPage/ListPage.jsx
@@ -12,10 +12,7 @@ const ListPage = () => {
       <div className={styles.mainContainer}>
         <h1 className={styles.helloCredit}>임시 크레딧</h1>
         <h2>요 왼쪽 부분에 후원을 기다리는 조공 텍스트(타이틀) 추가하기</h2>
-        {/* 후원을 기다리는 조공 텍스트는 TributeSlider 컴포넌트 안에서 구현하는게 좋아보이는데
-            코드를 조금 만져봤는데 제가 작성한 컴포넌트가 아니라서 헷갈리더라고요...
-            컴포넌트 코드 유지보수 측면에서도 해당 부분 작업하신 광조님이 추가해주시는게 좋아보입니다!
-        */}
+
         <TributeSlider />
         <ChartPage />
       </div>
