+++ src/styles/global.scss
@@ -1,7 +1,7 @@
 @import "./font";
 @import "./color";
 @import "../components/slider/slider.scss";
-@import "../components/loadingSpinner/loadingSpinner.scss";
+@import "../components/common/loadingSpinner.scss";
 
 *,
 *::before,
